refactor: Receipt의 패키지 변경
