+++ src/main/java/store/io/OutputView.java
@@ -3,8 +3,8 @@ package store.io;
 import java.util.List;
 import store.domain.MembershipManager;
 import store.domain.Product;
+import store.domain.Receipt;
 import store.domain.StoreHouse;
-import store.dto.Receipt;
 import store.io.parser.ProductListParser;
 import store.io.parser.ReceiptParser;
 import store.io.writer.Writer;
